trying file buffering to se if it goes faster

+++ www/recordings/management/commands/score_snippets.py
@@ -1,5 +1,7 @@
 import sys
 import time
+from io import BufferedReader
+from contextlib import closing
 
 from django.core.management.base import BaseCommand, CommandError
 from django.conf import settings
@@ -9,7 +11,7 @@ from kokako.detectors.intensity import Energy, LowEnergy, Amplitude
 
 from recordings.models import Score, Recording, Snippet, Detector
 
-
+BUFFER_SIZE = 1024*1024
 class Command(BaseCommand):
     def handle(self, *args, **options):
         detectors = [SimpleKiwi(), Energy(), LowEnergy(), Amplitude()]
@@ -28,25 +30,29 @@ class Command(BaseCommand):
         recordings = Recording.objects.all().order_by('?')
         for recording in recordings:
             snippets = Snippet.objects.filter(recording=recording).exclude(scores__detector=kiwi_detector).order_by('offset')
-            for snippet in snippets:
-                count = 0
-                for detector, db_detector in detectors:
-                    try:
-                        audio = Audio(snippet.get_audio(), snippet.recording.sample_rate)
-                        score = detector.score(audio)
-                        if not count:   
-                            print '%s %0.1f %0.1f' % (snippet, time.time() - now, score)
-                            now = time.time()
-                        try:
-                            s = Score.objects.get(detector=db_detector, snippet=snippet) 
-                            s.delete()
-                        except Score.DoesNotExist:
-                            pass
-                        s = Score(detector=db_detector, snippet=snippet,
-                            score=score)
-                        s.save()
-                    except KeyboardInterrupt:
-                        raise
-                    except:
-                        print detector, snippet, 'Scoring failed', sys.exc_info()[0]
-                    count += 1
+            if len(snippets):
+                fid = BufferedReader(open(recording.path, 'rb'), buffer_size=BUFFER_SIZE)
+                fid.peek(BUFFER_SIZE)
+                with closing(fid):
+                    for snippet in snippets:
+                        count = 0
+                        for detector, db_detector in detectors:
+                            try:
+                                audio = Audio(snippet.get_audio(fid), snippet.recording.sample_rate)
+                                score = detector.score(audio)
+                                if not count:   
+                                    print '%s %0.1f %0.1f' % (snippet, time.time() - now, score)
+                                    now = time.time()
+                                try:
+                                    s = Score.objects.get(detector=db_detector, snippet=snippet) 
+                                    s.delete()
+                                except Score.DoesNotExist:
+                                    pass
+                                s = Score(detector=db_detector, snippet=snippet,
+                                    score=score)
+                                s.save()
+                            except KeyboardInterrupt:
+                                raise
+                            except:
+                                print detector, snippet, 'Scoring failed', sys.exc_info()[0]
+                            count += 1
